Use YASnippet in Go mode and disable go-errcheck.
Flycheck was showing virtually everything as an error when you imported
a package.  Example

import "fmt"  <<  Would be an error as it couldn't find package.

Then whenever methods from that package were called they would also be
flagged as an error.  These were always due to the go-errcheck so
disabling it as it is just noise.

Also, use YASnippet when in Go mode.