Update Emacs safe local var and auth sources.
Auth sources needed updated so that it will read my custom authinfo
files tha include creds for both Gmail accounts.  It is not possible to
use just the default .authinfo as the hostnames in the file would be the
same since they are both Gmail accounts.

Allow auto-fill-mode to be set locally by adding it as a local variable
comment at the top of a file.  I use this in Org mode, in particular
when writing blog posts.