Removing Deoplete vim config.
Switched to maralla/completor.vim so the installation of deoplete had
been removed yet this config remained in place by accident.  It of
course causes an issue when Deoplete is not installed, I'd not realised
this on other machines as Deoplete install must've still be hanging
around (PlugClean had not been run).