Autostart things with AwesomeWM.
On Fedora since it is using Gnome and friends inc. GDM the .xinitrc
doesn't seem to be respected so start stuff I'd normally start via
.xinitrc using Awesome's own method.  The shell script _should_ check to
make sure if a process is running already so this hopefully won't
interfere when used on more minimal systems where the .xinitrc is being
respected such as BSD and Voiid.