Make most Org mode captures go to INBOX.
The inbox is a special heading within the org files as it is intended as
a temporary area for captured things to go into with a view to being
refiled in a more suitable place later.  To reflext this make the
heading all in uppercase to make it stand out i.e. INBOX.