Fix git completion on Mac OS
Update to the latest git-completion.bash that the zsh one uses.

The zsh completion script seems to look for the bash script in various
locations, the first of which is the same dir that it is in. It
mustn't have been finding the bash one where I had it in
`~/.zsh/lib/git-completion.bash` and rather than modify the script,
changing it from upstream, just place it alongside to fix completion
on a Mac.  I think I've not noticed this elsewhere because the script
gets installed along with packages when installing git and those
common locations are also looked in, the Git install on Mac however
uses a strange location and is therefore not being picked up.
Besides, I wanna use my 'vendored' version within my dotfiles.