Add Victor Mono font, make it default in Emacs. Cleanup fonts
Remove erroneous fonts such as the Gohu tarball and Proggy ttf, Proggy
is only gonna look right as a bitmap font and it is just too small for
modern screen resolutions.

Set Victor Mono font as default in Emacs via a slightly different way
as with `set-frame-font` it was only working for 'normal' mode Emacs
on Mac OS, it wasn't getting set when using emacsclient.