Simplify RUSTC_WRAPPER sccache binary, don't need full path
Having the full path here causes an error if sccache has been
installed via Nix instead of Cargo.  I think originally this HAD to be
the full path but that seems to be no longer the case.  I wonder if
this can therefore be moved back out into the Cargo config instead of
an env var here.  I'll test this at a later date.