Set sccache as rustc wrapper via Cargo config
This used to seem to NEED to be an absolute path which meant it broke
between Mac and Linux, you also couldn't use $HOME and it also meant
that there were issues depending on if sccache was installed via Cargo
or via Nix.

I THINK this might now work with just the binary name as long as it is
in $PATH so it is cleaner to put back in Cargo config rather than yet
another env var.