fix(emacs): set folding symbol to one that ETBembo has
Not sure why but the downwards arrow symbol shows as a Chinese symbol
on new laptop.  It turns out ETBembo which is set for headings in Org
mode doesn't have that symbol at all, not sure how this ever worked on
work machine but this puts it to a symbol that ETBembo does have and
should show fine.

Also remove org-habit as I never use it and am stripping back unused stuff.