fix(emacs): cloning of my org-cv fork on sourcehut
Elpaca understands 'sourcehut' as a keyword and knows from that what
URL it needs to hit to clone the repo.  It also knows that Sourcehut
repos have the form '~rosstimson/org-cv' with the preceding '~' and
sorts that automatically.

Without this fix Elpaca was failing to clone and seemed to be
expecting the repo within my home directory, presumably getting
confused by the '~'.