fix(emacs:) remove treesit-auto
Switching to the treesit version of modes seems to be causing issues
that were tricky to track down, especially with Go.  Rather than just
exclude certain langs/modes just ditch Treesit-Auto and explicitly set
when I wanna use the Treesit versions.  It was always known that the
treesit stuff was very new and not fully ready from primetime yet.