refactor(restic): tweak scheduling for resticprofile
I think it's preventing my laptop from sleeping completely rather than
it being just for when backups are running which is what I had
assumed.

Also, when scheduled for `weekly` it's trying to do backups just after
midnight on Monday mornings, it always fails cos I'm not around to
auth with 1Password so set the schedule to every Monday but at a time
I'm more likely to be around.