fix(shell): getting rclone config password
Having it as an env var has turned out to be a PITA and it wants me to
use fingerprint for every new terminal window as well as when
launching Emacs.  There is probably a way around that but I never
really wanted it in the env anyway, I wanted it to just prompt
whenever the rclone command is used.  By switching this to an alias
which has the `--password-command` in there I can do just that, it'll
only prompt for fingerprint via 1Password when the `rclone` command is used.