refactor(emacs): go-ts-mode indent size & grammar settings
Put the go-ts-mode treesit lang back into the
treesit-language-source-alist which I'd already set rather than having
it duplicated within the :init of the package itself.

Also set the indent-offset to 4, the default was 8.  When set to 8 it
was adding an extra tab as my default tab-width is 4.  This feels like
a bug, it shouldn't matter what size you choose to represent a tab
with but for some reason an extra tab was getting added, this allow me
to still use a tab-width of 4 but avoid ending up with go-ts-mode
adding an extra tab meaning indents were 2 tabs.